settings
git-svn-id: https://plugins.svn.wordpress.org/squirrly-seo/trunk@680287 b8457f37-d9ea-0310-8a92-e5e31aec5664

+++ controllers/SQ_Menu.php
@@ -61,7 +61,7 @@ class SQ_Menu extends SQ_FrontController {
                 
           switch (SQ_Tools::getValue('action')){
             case 'sq_howto':
-                SQ_Tools::saveOptions('sq_howto', SQ_Tools::getValue('sq_howto'));
+                SQ_Tools::saveOptions('sq_howto', (int)SQ_Tools::getValue('sq_howto'));
                 exit();
                 break;
             case 'sq_update':
